+++ CHANGELOG.rst
@@ -1,6 +1,20 @@
 Changelog
 =========
 
+0.0.2 - 2014-11-20
+------------------
+
+* Adds resources for
+  `AWS CloudFormation <http://aws.amazon.com/cloudformation/>`_ and
+  `AWS OpsWorks <http://aws.amazon.com/opsworks/>`_.
+* Update to Botocore 0.73.0 and JMESPath 0.5.0
+* Adds support for
+  `AWS CodeDeploy <http://aws.amazon.com/codedeploy/>`_,
+  `AWS Config <http://aws.amazon.com/config/>`_,
+  `AWS KMS <http://aws.amazon.com/kms/>`_,
+  `AWS Lambda <http://aws.amazon.com/lambda/>`_.
+* Make requests with a customized HTTP user-agent
+
 0.0.1 - 2014-11-11
 ------------------